refactor: split tab and history api clients
